% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bettingapir-package.R
\docType{package}
\name{bettingapir-package}
\alias{bettingapir}
\alias{bettingapir-package}
\title{`{bettingapir}`: An Interface to Betting API for R}
\description{
`{bettingapir}` provides users a range of functions designed to communicate with Betting API
from R.
}
