% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bettingapir_auth.R
\name{bettingapir_auth}
\alias{bettingapir_auth}
\title{Set Betting API's Credentials}
\usage{
bettingapir_auth(token = Sys.getenv("BETTINGAPI_TOKEN"))
}
\arguments{
\item{token}{Betting API's token.}
}
\description{
Set Betting API's Credentials
}
